# SIMD optimisation in Hudwarp code