--- a/bmedll/MiscRenderHooks.cpp
+++ b/bmedll/MiscRenderHooks.cpp
@@ -6,98 +6,6 @@
 
 #include <d3d11_1.h>
 
-bool shouldUseGPUHudwarp = false;
-bool isHudwarpDisabled = false;
-
-bool isRenderingHud = false;
-HudwarpProcess* hudwarpProcess = nullptr;
-
-typedef void(__fastcall* sub_18000BAC0_type)(float*, float*, float*);
-sub_18000BAC0_type sub_18000BAC0_org = nullptr;
-void __fastcall sub_18000BAC0(float* a1, float* a2, float* a3)
-{
-    // Ported from TFORevive by Barnaby
-
-    if (!shouldUseGPUHudwarp)
-    {
-        // If hudwarp is disabled and running on CPU do just the scaling
-        if (isHudwarpDisabled)
-        {
-            // Still perform scaling for HUD when warping is disabled
-            float viewWidth = a1[2];
-            float viewHeight = a1[3];
-
-            float xScale = a1[7];
-            float yScale = a1[9];
-
-            a3[0] = (a2[0] - 0.5f * viewWidth) * xScale + 0.5f * viewWidth;
-            a3[1] = (a2[1] - 0.5f * viewHeight) * yScale + 0.5f * viewHeight;
-            a3[2] = a2[2];
-            return;
-        }
-
-        return sub_18000BAC0_org(a1, a2, a3);
-    }
-
-    // We prevent the hud from reaching bounds of render texture by adding a border of 0.025 * width/height to the texture
-    // We need to offset the verts to account for that here
-    const float hudOffset = HUD_TEX_BORDER_SIZE;
-    const float hudScale = 1.0f - 2.0f * hudOffset;
-
-    float viewWidth = a1[2];
-    float viewHeight = a1[3];
-
-    a3[0] = a2[0] * hudScale + hudOffset * viewWidth;
-    a3[1] = a2[1] * hudScale + hudOffset * viewHeight;
-    a3[2] = a2[2];
-}
-
-typedef void(__fastcall* CMatSystemSurface__ApplyHudwarpSettings_type)(void*, HudwarpSettings*, unsigned int, unsigned int);
-CMatSystemSurface__ApplyHudwarpSettings_type CMatSystemSurface__ApplyHudwarpSettings_org = nullptr;
-void __fastcall CMatSystemSurface__ApplyHudwarpSettings(void* thisptr, HudwarpSettings* hudwarpSettings, unsigned int screenX, unsigned int screenY)
-{
-    // Ported from TFORevive by Barnaby
-
-    static ConVarRef hudwarp_chopsize{ "hudwarp_chopsize" };
-    unsigned int originalChopsize = hudwarp_chopsize->GetInt();
-
-    static ConVarRef hudwarp_disable{ "hudwarp_disable" };
-    HudwarpSettings newSettings = *hudwarpSettings;
-    if (hudwarp_disable->GetInt())
-    {
-        // Override hudwarp settings if hudwarp_disable is 1.
-        // NOTE: Comment below refers to original CPU version, we can set them to 0 when using our shader.
-        // Stuff breaks if you set the warp values to 0.
-        // Respawn set them to a min of 1deg in rads (0.017453292), we can do that too because it'll result in so little distortion you won't notice it :)
-        newSettings.xWarp = 0.0f;
-        if (newSettings.xScale > 1.0f) newSettings.xScale = 1.0f;
-        newSettings.yWarp = 0.0f;
-        if (newSettings.yScale > 1.0f) newSettings.yScale = 1.0f;
-    }
-
-    if (hudwarpProcess)
-        hudwarpProcess->UpdateSettings(&newSettings);
-
-    static ConVarRef hudwarp_use_gpu{ "hudwarp_use_gpu" };
-    // If using GPU hudwarp or hudwarp is disabled do this
-    // Replace chopsize, it gets set from the cvar in CMatSystemSurface__ApplyHudwarpSettings
-    if (hudwarp_use_gpu->GetInt() || hudwarp_disable->GetInt())
-    {
-        if (screenX > screenY) [[likely]]
-        {
-            hudwarp_chopsize->SetValueInt(screenX);
-        }
-        else
-        {
-            hudwarp_chopsize->SetValueInt(screenY);
-        }
-    }
-
-    CMatSystemSurface__ApplyHudwarpSettings_org(thisptr, &newSettings, screenX, screenY);
-
-    hudwarp_chopsize->SetValueInt(originalChopsize);
-}
-
 ID3D11Device* pDevice;
 ID3D11DeviceContext** ppID3D11DeviceContext;
 ID3DUserDefinedAnnotation* pPerf;
@@ -132,6 +40,13 @@ void QueueEndEvent()
     SetPixMarker(queuedRenderContext, 222222, ""); // hijack color field for our purposes
 }
 
+bool isRenderingHud = false;
+
+bool shouldUseGPUHudwarp = false;
+bool isHudwarpDisabled = false;
+
+HudwarpProcess* hudwarpProcess = nullptr;
+
 void HudRenderStart()
 {
     if (!shouldUseGPUHudwarp)
@@ -148,16 +63,6 @@ void HudRenderFinish()
 
     if (hudwarpProcess)
         hudwarpProcess->Finish();
-}
-
-typedef void(__fastcall* OnWindowSizeChanged_type)(unsigned int, unsigned int, bool);
-OnWindowSizeChanged_type OnWindowSizeChanged_org = nullptr;
-void __fastcall OnWindowSizeChanged(unsigned int w, unsigned int h, bool isInGame)
-{
-    OnWindowSizeChanged_org(w, h, isInGame);
-
-    if (hudwarpProcess)
-        hudwarpProcess->Resize(w, h);
 }
 
 HookedFuncStaticWithType<SetPixMarker_type> sub_5ADC0("materialsystem_dx11.dll", 0x5ADC0);
@@ -213,6 +118,103 @@ void __fastcall RenderHud_Hook(__int64 a1, __int64 a2, __int64 a3)
     QueueBeginEvent("HUD");
     RenderHud(a1, a2, a3);
     QueueEndEvent();
+}
+
+typedef void(__fastcall* sub_18000BAC0_type)(float*, float*, float*);
+sub_18000BAC0_type sub_18000BAC0_org = nullptr;
+void __fastcall sub_18000BAC0(float* a1, float* a2, float* a3)
+{
+    // Ported from TFORevive by Barnaby
+
+    if (!shouldUseGPUHudwarp)
+    {
+        // If hudwarp is disabled and running on CPU do just the scaling
+        if (isHudwarpDisabled)
+        {
+            // Still perform scaling for HUD when warping is disabled
+            float viewWidth = a1[2];
+            float viewHeight = a1[3];
+
+            float xScale = a1[7];
+            float yScale = a1[9];
+
+            a3[0] = (a2[0] - 0.5f * viewWidth) * xScale + 0.5f * viewWidth;
+            a3[1] = (a2[1] - 0.5f * viewHeight) * yScale + 0.5f * viewHeight;
+            a3[2] = a2[2];
+            return;
+        }
+
+        return sub_18000BAC0_org(a1, a2, a3);
+    }
+
+    // We prevent the hud from reaching bounds of render texture by adding a border of 0.025 * width/height to the texture
+    // We need to offset the verts to account for that here
+    const float hudOffset = HUD_TEX_BORDER_SIZE;
+    const float hudScale = 1.0f - 2.0f * hudOffset;
+
+    float viewWidth = a1[2];
+    float viewHeight = a1[3];
+
+    a3[0] = a2[0] * hudScale + hudOffset * viewWidth;
+    a3[1] = a2[1] * hudScale + hudOffset * viewHeight;
+    a3[2] = a2[2];
+}
+
+typedef void(__fastcall* OnWindowSizeChanged_type)(unsigned int, unsigned int, bool);
+OnWindowSizeChanged_type OnWindowSizeChanged_org = nullptr;
+void __fastcall OnWindowSizeChanged(unsigned int w, unsigned int h, bool isInGame)
+{
+    OnWindowSizeChanged_org(w, h, isInGame);
+
+    if (hudwarpProcess)
+        hudwarpProcess->Resize(w, h);
+}
+
+typedef void(__fastcall* CMatSystemSurface__ApplyHudwarpSettings_type)(void*, HudwarpSettings*, unsigned int, unsigned int);
+CMatSystemSurface__ApplyHudwarpSettings_type CMatSystemSurface__ApplyHudwarpSettings_org = nullptr;
+void __fastcall CMatSystemSurface__ApplyHudwarpSettings(void* thisptr, HudwarpSettings* hudwarpSettings, unsigned int screenX, unsigned int screenY)
+{
+    // Ported from TFORevive by Barnaby
+
+    static ConVarRef hudwarp_chopsize{ "hudwarp_chopsize" };
+    unsigned int originalChopsize = hudwarp_chopsize->GetInt();
+
+    static ConVarRef hudwarp_disable{ "hudwarp_disable" };
+
+    HudwarpSettings newSettings = *hudwarpSettings;
+    if (hudwarp_disable->GetInt())
+    {
+        // Override hudwarp settings if hudwarp_disable is 1.
+        // NOTE: Comment below refers to original CPU version, we can set them to 0 when using our shader.
+        // Stuff breaks if you set the warp values to 0.
+        // Respawn set them to a min of 1deg in rads (0.017453292), we can do that too because it'll result in so little distortion you won't notice it :)
+        newSettings.xWarp = 0.0f;
+        if (newSettings.xScale > 1.0f) newSettings.xScale = 1.0f;
+        newSettings.yWarp = 0.0f;
+        if (newSettings.yScale > 1.0f) newSettings.yScale = 1.0f;
+    }
+
+    if (hudwarpProcess)
+        hudwarpProcess->UpdateSettings(&newSettings);
+
+    static ConVarRef hudwarp_use_gpu{ "hudwarp_use_gpu" };
+    // If using GPU hudwarp or hudwarp is disabled do this
+    // Replace chopsize, it gets set from the cvar in CMatSystemSurface__ApplyHudwarpSettings
+    if (hudwarp_use_gpu->GetInt() || hudwarp_disable->GetInt())
+    {
+        if (screenX > screenY) [[likely]]
+        {
+            hudwarp_chopsize->SetValueInt(screenX);
+        }
+        else
+        {
+            hudwarp_chopsize->SetValueInt(screenY);
+        }
+    }
+
+    CMatSystemSurface__ApplyHudwarpSettings_org(thisptr, &newSettings, screenX, screenY);
+
+    hudwarp_chopsize->SetValueInt(originalChopsize);
 }
 
 void SetupHudwarp()