Ported over from TFORevive, brings similar performance improvements

Tested before and after with max settings apart from AA (2x MSAA, higher settings tank performance regardless) with my RX 7700XT + Ryzen 5 5600X PC
Before: ~360fps average, ~240fps min
After: ~420fps average, ~360fps min